refactor(Amenities): allow vertical spacing to be optional
